# artifacts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tasks/attachments/staged": {
      "post": {
        "operationId": "TasksController_stageAttachment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
...
    "/api/v1/tasks/{id}/attachments": {
      "post": {
        "operationId": "TasksController_uploadAttachment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
...
  },
  "info": {
    "title": "LankaDWS API",
    "description": "Permission-scoped LankaDWS modular monolith API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "lankadws_session"
      }
    },
...
  }
}